Refactored code to run/show platform specific commands and messages
